Fix exception swallowing in CLI commands.
The CLI commands were using a try/catch block to determine if the
incoming dataset identifier was a URI or a name. If the URI was a
dataset URI but another operation failed with IllegalArgumentException,
for example a problem during creation, then the catch would retry the
operation assuming the input was a dataset name.

+++ kite-tools/src/main/java/org/kitesdk/cli/commands/DeleteDatasetCommand.java
@@ -43,10 +43,9 @@ public class DeleteDatasetCommand extends BaseDatasetCommand {
     }
 
     for (String datasetUriOrName : datasets) {
-      try {
+      if (isDataUri(datasets.get(0))) {
         Datasets.delete(datasetUriOrName);
-      } catch (IllegalArgumentException _) {
-        // Not a "dataset" or "view" URI, use the default repository
+      } else {
         repo.delete(datasetUriOrName);
       }
       console.debug("Deleted dataset {}", datasetUriOrName);
